// src/types/CommentClient.ts
/*
 * Copyright 2023 WPPConnect Team
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/** Optional until the matching WPPConnect/WA-JS changes are released. */
export interface CommentClient {
  getComments?: (messageId: string) => Promise<unknown[]>;
  sendCommentMessage?: (
    messageId: string,
    text: string
  ) => Promise<{
    messageSendResult: string;
    t?: number;
    count?: number | null;
  }>;
  onCommentMessage?: (callback: (event: unknown) => void) => unknown;
}